Remove demo link.
The demo is too much work for me to maintain. I don't want to give
anyone admin access, and I don't want to enable email for fear of
spammers using the site for evil. These two things severely limit the
functionality and people are confused about how to do a lot of things
that would be clear if they were getting emails and/or had admin access.

I think a slick demo video would be a better idea. I'll start on that
soon.